+++ cmd/hidden_lake/traffic/internal/handler/handler_test.go
@@ -161,7 +161,6 @@ func testNewClient() client.IClient {
 	return client.NewClient(
 		message.NewSettings(&message.SSettings{
 			FMessageSizeBytes: testutils.TCMessageSize,
-			FWorkSizeBits:     testutils.TCWorkSize,
 		}),
 		privKey,
 	)