+++ pkg/client/message/settings.go
@@ -6,13 +6,11 @@ var (
 
 type SSettings sSettings
 type sSettings struct {
-	FWorkSizeBits     uint64
 	FMessageSizeBytes uint64
 }
 
 func NewSettings(pSett *SSettings) ISettings {
 	return (&sSettings{
-		FWorkSizeBits:     pSett.FWorkSizeBits,
 		FMessageSizeBytes: pSett.FMessageSizeBytes,
 	}).mustNotNull()
 }
@@ -22,10 +20,6 @@ func (p *sSettings) mustNotNull() ISettings {
 		panic(`p.FMessageSizeBytes == 0`)
 	}
 	return p
-}
-
-func (p *sSettings) GetWorkSizeBits() uint64 {
-	return p.FWorkSizeBits
 }
 
 func (p *sSettings) GetMessageSizeBytes() uint64 {