+++ pkg/client/queue/examples/main.go
@@ -25,7 +25,6 @@ func main() {
 		client.NewClient(
 			message.NewSettings(&message.SSettings{
 				FMessageSizeBytes: (1 << 12),
-				FWorkSizeBits:     10,
 			}),
 			asymmetric.NewRSAPrivKey(1024),
 		),